+++ routers/goodses.js
@@ -13,9 +13,17 @@
 
 const [{resolve, join}, {createReadStream, createWriteStream}, parse] = [require("path"), require(("fs")), require("co-busboy")];
 const services = resolve(__dirname, "..", "services");
-const [{getGoodsList, getGoodsDetailed, increaseCount, checkGoodsExist}, {getGoodsFileInfo}, {getWorkflowNode, updateNodeProduceFile}, {createGoods, getGoodesHouseAddress},
-  {throwLackParameters}] = [require(resolve(services, "goodsService")), require(resolve(services, "fileSystem")), require(resolve(services, "workflowService")),
-  require(resolve(services, "goodsService")), require(resolve(__dirname, "..", "errors"))];
+const [
+  {getGoodsList, getGoodsDetailed, increaseCount, checkGoodsExist, createGoods, getGoodesHouseAddress},
+  {getGoodsFileInfo},
+  {getWorkflowNode, updateNodeProduceFile, obmitUploadFileAuthorize},
+  {throwLackParameters}
+] = [
+  require(resolve(services, "goodsService")),
+  require(resolve(services, "fileSystem")),
+  require(resolve(services, "workflowService")),
+  require(resolve(__dirname, "..", "errors"))
+];
 
 const list = function* (next) {
   if (this.error) {
@@ -62,6 +70,7 @@ const updateNode = function* (next) {
   try {
     const [fields, workflowNodeId, authorized] = [parse(this.req), this.params.nodeId, this.authorized];
     const _ = yield getWorkflowNode(workflowNodeId);
+    yield obmitUploadFileAuthorize(_.workflow, authorized);
 
     if (!_) {
       throwLackParameters();